MDL-62194 enrol: Stop erratic behaviour on manual enrolment
Before this change when searching for a user on the manual
enrolment form AJAX requests would be sent off whenever the
user types a character into the search box (subject to browser
throttling which could vary). If a Moodle site has many users
the earlier queries could return after later queries had completed.
This resulted in the user list seeming to filter down and then suddenly
showing lots of users.

This change does the following:

1) First it stops a request being sent to Moodle until there is a pause
of 500ms in a user's typing. This should mean that the initial request sent
will usually return fewer results and so take less time.

2) If a request is sent it will block new requests being created until it
has returned. This will mean that an earlier request can never overwrite
a later request.

3) It remembers the latest query made while new AJAX requests are blocked
and sends that value automatically after the active request completes. This
means that if a request is started and the user continues to type the list of
users will eventually be filtered to it.

All of this should mean that searching for a user to enrol sends less requests to
Moodle, but accurately displays a filtered list of possible users.