MDL-64550 course: Activity navigation links should wrap
The btn class stops white space from wrapping, so if an activity
name was long it could not wrap and would overlap the activity jump
menu and possibly the other activity names.

The btn class appear to have been used to try to align the links
with the jump menu, so py-2 has been used to try to replicate that to
without the negative side affects.