MDL-71073 courses: Stop memory errors when building caches
When running core_course_search_courses on a site with a large
number of courses PHP could run out of memory. It looks like it was
because a lot of unnecessary data was being retrieved just so we can
store an ordered list of course ids.

This change should greatly reduce the amount of data retrieved to
build the cache entries.